Get rid of all the Wconversion warnings

+++ 2D/Ising-bis.cpp
@@ -12,7 +12,7 @@ namespace vb {
 
 class Ising : public Bitmap<int> {
 public:
-    Ising(const Hub &H, size_t n, double beta, int con) : Bitmap<int>(H.title, {n, n}), con(con), p(2 * con + 1) {
+    Ising(const Hub &H, size_t n, double beta, int con) : Bitmap<int>(H.title, {n, n}), con(con), p(2 * size_t(con) + 1) {
         for (auto z : coo_range(size)) put(z, prng.bernoulli(.5) ? 1 : -1);
         if (con == 6) {
             size_t m = 2 * n / 3;
@@ -20,7 +20,7 @@ public:
                 for (size_t j = 0; j < n; ++j)
                     if ((j > m) || (i < j / 2) || (i > m + j / 2)) put({i, j}, 0);
         }
-        for (int c = -con; c <= con; ++c) p[c + con] = 1 / (1 + exp(-beta * c));
+        for (int c = -con; c <= con; ++c) p[size_t(c + con)] = 1 / (1 + exp(-beta * c));
     };
 
     void run() {
@@ -29,7 +29,7 @@ public:
             if (at(z) == 0) continue;
             int c = 0;
             for (int d = 0; d < con; ++d) c += atp(z + dz[d]);
-            put(z, prng.bernoulli(p[c + con]) ? 1 : -1);
+            put(z, prng.bernoulli(p[size_t(c + con)]) ? 1 : -1);
         }
     }
